requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.6.0
google-genai==0.2.1
python-dotenv==1.0.1
